training_code/requirements.txt
torch>=2.8.0
torchvision>=0.23.0
numpy>=1.23
Pillow>=10.0.0
PyYAML>=6.0
tqdm>=4.66
dists-pytorch>=0.1.1
